re-organised into subsystems & added res cancellation
